Covert Integer to ASCII
Convert Integer to ASCII representation.
Given a 16 bit integer, write a function to return a string of
characters denoting that integer (similar to sprintf())
e.g. if num = 0xBAD1, your string should be "BAD1", else if
num = 123, your string should be "0123". remember that string will
require atleast a fixed size of 5 chars: 4 for the number + 1 for '\0'
in the end